# src/main/resources/static/docs/openapi3.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Image Processing API",
    "description" : "이미지 처리 위임 서버 API. 접수/조회/목록.",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ],
  "tags" : [ ],
  "paths" : {
    "/api/jobs" : {
      "get" : {
        "tags" : [ "Jobs" ],
        "summary" : "작업 목록 조회",
        "description" : "작업 목록을 상태 필터·페이징으로 조회한다.",
        "operationId" : "jobs-list",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "description" : "상태 필터(선택)",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "페이지(0-base, 기본 0)",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "페이지 크기(기본 20)",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-jobs-937641076"
                },
                "examples" : {
                  "jobs-list" : {
                    "value" : "{\"jobs\":[{\"trackingId\":\"aa702b21-936d-471d-b907-5d171dab1a2c\",\"status\":\"PROCESSING\",\"imageUrl\":\"https://img/1.png\",\"result\":null,\"errorCode\":null,\"retryCount\":1,\"createdAt\":\"2026-01-01T00:00:00Z\",\"updatedAt\":\"2026-01-01T00:00:05Z\"}],\"page\":0,\"size\":20,\"totalElements\":1,\"totalPages\":1}"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Jobs" ],
        "summary" : "이미지 처리 접수",
        "description" : "이미지 처리 작업을 접수하고 추적 식별자를 반환한다. 신규는 202, 멱등 키로 수렴된 중복은 200.",
        "operationId" : "jobs-submit",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "description" : "멱등 키(선택). 없으면 imageUrl 해시로 대체",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json;charset=UTF-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/api-jobs811708560"
              },
              "examples" : {
                "jobs-submit" : {
                  "value" : "{\"imageUrl\":\"https://img/1.png\"}"
                }
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "description" : "202",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-jobs-939044843"
                },
                "examples" : {
                  "jobs-submit" : {
                    "value" : "{\"trackingId\":\"c3436024-6b63-4cab-8b0d-fa30de564bb7\",\"status\":\"PENDING\"}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{trackingId}" : {
      "get" : {
        "tags" : [ "Jobs" ],
        "summary" : "단일 작업 조회",
        "description" : "trackingId로 작업의 진행 상태/결과/실패 사유를 조회한다.",
        "operationId" : "jobs-get",
        "parameters" : [ {
          "name" : "trackingId",
          "in" : "path",
          "description" : "작업 추적 식별자(UUID)",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-jobs-trackingId-1717844821"
                },
                "examples" : {
                  "jobs-get" : {
                    "value" : "{\"trackingId\":\"eaad1e50-bbf8-446d-a424-d617aca90cc8\",\"status\":\"COMPLETED\",\"imageUrl\":\"https://img/1.png\",\"result\":\"result-data\",\"errorCode\":null,\"retryCount\":0,\"createdAt\":\"2026-01-01T00:00:00Z\",\"updatedAt\":\"2026-01-01T00:00:10Z\"}"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "api-jobs-937641076" : {
        "required" : [ "page", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "size" : {
            "type" : "number",
            "description" : "페이지 크기"
          },
          "jobs" : {
            "type" : "array",
            "items" : {
              "required" : [ "createdAt", "imageUrl", "retryCount", "status", "trackingId", "updatedAt" ],
              "type" : "object",
              "properties" : {
                "createdAt" : {
                  "type" : "string",
                  "description" : "생성 시각"
                },
                "retryCount" : {
                  "type" : "number",
                  "description" : "재시도 누적 횟수"
                },
                "imageUrl" : {
                  "type" : "string",
                  "description" : "처리 대상 이미지 URL"
                },
                "updatedAt" : {
                  "type" : "string",
                  "description" : "최종 갱신 시각"
                },
                "status" : {
                  "type" : "string",
                  "description" : "작업 상태"
                },
                "trackingId" : {
                  "type" : "string",
                  "description" : "작업 추적 식별자"
                }
              }
            }
          },
          "totalPages" : {
            "type" : "number",
            "description" : "전체 페이지 수"
          },
          "page" : {
            "type" : "number",
            "description" : "현재 페이지"
          },
          "totalElements" : {
            "type" : "number",
            "description" : "전체 작업 수"
          }
        }
      },
      "api-jobs-trackingId-1717844821" : {
        "required" : [ "createdAt", "imageUrl", "retryCount", "status", "trackingId", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "string",
            "description" : "성공 시 결과 문자열",
            "nullable" : true
          },
          "createdAt" : {
            "type" : "string",
            "description" : "생성 시각"
          },
          "retryCount" : {
            "type" : "number",
            "description" : "재시도 누적 횟수"
          },
          "imageUrl" : {
            "type" : "string",
            "description" : "처리 대상 이미지 URL"
          },
          "updatedAt" : {
            "type" : "string",
            "description" : "최종 갱신 시각"
          },
          "status" : {
            "type" : "string",
            "description" : "작업 상태"
          },
          "trackingId" : {
            "type" : "string",
            "description" : "작업 추적 식별자"
          }
        }
      },
      "api-jobs811708560" : {
        "required" : [ "imageUrl" ],
        "type" : "object",
        "properties" : {
          "imageUrl" : {
            "type" : "string",
            "description" : "처리할 이미지 URL"
          }
        }
      },
      "api-jobs-939044843" : {
        "required" : [ "status", "trackingId" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "description" : "작업 상태"
          },
          "trackingId" : {
            "type" : "string",
            "description" : "작업 추적 식별자(UUID)"
          }
        }
      }
    }
  }
}